BACKEND/requirements.txt
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
django-simple-history==3.10.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
psycopg2
PyJWT==2.10.1
python-decouple==3.8
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
celery==5.3.4
reportlab==4.4.9
django-filter==25.1
Pillow==12.1.0
